backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
pymongo==4.7.3
PyJWT==2.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
requests==2.32.5
google-generativeai==0.8.3
Pillow==11.0.0
